added localized exception to root page delete
root page can't be deleted in the way it is implemented,
so for now a localized exception is throws

for the future I suggest we handle such action in the "gear"-actions
that are to be implemented, @see also issue #7

+++ modules/widget/page_detail/PageDetailWidgetModule.php
@@ -165,6 +165,9 @@ class PageDetailWidgetModule extends PersistentWidgetModule {
 	
 	public function deletePage() {
 		$oPage = PageQuery::create()->findPk($this->iPageId);
+		if($oPage->isRoot()) {
+			throw new LocalizedException('wns.page.delete_root_restriction_message', array('page_name' => $oPage->getName()));
+		} 
 		$oPage->delete();
 		return $this->iPageId;
 	}
